+++ .gitignore
@@ -1,4 +1,4 @@
-/secrets.py
+/app/secrets.py
 ### Example user template template
 ### Example user template
 
@@ -34,7 +34,7 @@ var/
 *.egg
 
 # PyInstaller
-#  Usually these files are written by a python script from a template
+#  Usually these files are written by a python script from a templates
 #  before PyInstaller builds the exe, so as to inject date/other infos into it.
 *.manifest
 *.spec
