+++ app/templates/index.html
@@ -1,0 +1,79 @@
+<!DOCTYPE html>
+
+
+<html lang="en">
+
+
+<head>
+
+    <meta charset="UTF-8">
+    <title>pyGlialo - giving away the goodies</title>
+    <link rel="stylesheet" href="static/bootstrap.min.css">
+    <link rel="stylesheet" href="static/bootstrap-theme.min.css">
+    <link rel="stylesheet" href="static/jumbotron-narrow.css">
+</head>
+<body>
+<div class="container">
+    <div class="header clearfix">
+        <h2 class="text-muted">pyGlialo</h2>
+        <i> giving away the goodies</i>
+    </div>
+
+    <div class="jumbotron">
+
+        <p class="lead">Rolling for Jetbrains 3</p>
+
+        <h1>{{ winner.name }}</h1>
+
+        <p class="lead"><a href="#">Meetup Profile</a></p>
+
+        <p>
+            <a class="btn btn-lg btn-success" href="save/{{winner.name}}" role="button">Claim the prize!</a>
+            <a class="btn btn-lg btn-warning" href="/" role="button">Another one</a>
+            <a class="btn btn-lg btn-danger" href="#" role="button">Leave slot empty</a>
+        </p>
+
+        <p><a class="btn btn-lg btn-primary" href="#" role="button">Finalize</a></p>
+    </div>
+
+    <div class="row marketing">
+        <div class="col-lg-6">
+            <h4>Jetbrains 1</h4>
+
+            <p>Giuseppe A.</p>
+
+            <h4>Jetbrains 2</h4>
+
+            <p>Francesco B.</p>
+
+            <h4>Jetbrains 3</h4>
+
+            <p><i>Still Rolling</i></p>
+        </div>
+
+        <div class="col-lg-6">
+            <h4>Elastic 1</h4>
+
+            <del>This slot is empty</del>
+
+            <h4>Elastic 2</h4>
+
+            <p>Waiting for Jetbrains 3</p>
+
+            <h4>Elastic 3</h4>
+
+            <p>Waiting for Jetbrains 3</p>
+        </div>
+    </div>
+
+    <footer class="footer">
+        <p>&copy; python milano - pyGlialo</p>
+    </footer>
+
+</div> <!-- /container -->
+
+<script src="static/jquery.min.js"></script>
+<script src="static/bootstrap.min.js"></script>
+
+</body>
+</html>
